update hook, fix required params

+++ pre-commit.sh
@@ -4,6 +4,9 @@ set -ex
 yarn pretty-quick --staged
 rm -rf lib
 yarn build
-rm -rf node_modules
-yarn install --production
-git add lib node_modules
+git add lib
+if ! git diff --name-only --cached --exit-code package.json yarn.lock; then
+  rm -rf node_modules
+  yarn install --production
+  git add node_modules
+fi
